Use imagej.net URL for all developers
And remove the other info, which is prone to obsolescence.

This information is available from imagej.net user profiles,
in an indirect way which can be updated in case it changes,
rather than being baked into the POM forever.

This change will make automatic generation of informational
sidebars on the wiki easier to accomplish.