Update dependency versions
This updates to pom-scijava 15.0.0 and unpins relevant versions.

It also updates OpenCSV from au.com.bytecode:opencsv:2.4
to com.opencsv:opencsv:3.9, in preparation for pom-scijava 16.

And it updates from ome:formats-common to org.openmicroscopy:ome-common.

Lastly, courtesy of @bogovicj, it makes the necessary changes to
WarpedSource to address backwards-incompatible API changes in the BDV.